Move to flow control to the FrameWriter instead - We only want to throttle data frames
