+++ src/Servers/Kestrel/Core/src/Internal/Http2/Http2OutputProducer.cs
@@ -11,7 +11,6 @@ using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2.FlowControl;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure.PipeWriterHelpers;
-using Microsoft.Extensions.Logging;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2;
 
@@ -149,7 +148,6 @@ internal class Http2OutputProducer : IHttpOutputProducer, IHttpOutputAborter, IV
             return wasEmpty && _unconsumedBytes > 0;
         }
     }
-
 
     public void StreamReset()
     {
